tests: split AssignmentHelpersTests.swift (1,114 lines → 3 files)
Follow-on to #490.  Last of the four oversized test files.

Layout:
  AssignmentHelpersTestCase.swift     — base + makeFile/makeZip/notebookData
  AssignmentHelpersManifestTests.swift — path sanitization, manifest ops,
                                         buildSuiteEntries, requirement
                                         detection (15)
  AssignmentHelpersUtilityTests.swift  — grade percent, CSV, dates, edit-suite
                                         resolution, merge, practiceLab (19)

34 tests preserved; all pass under --parallel.  #490 is now complete —
four mega-classes split into 15 focused subclasses inheriting from four
new TestCase base classes.
